+++ packages/server/src/dal/in-memory/repositories/in-memory-article-repository.ts
@@ -3,11 +3,14 @@ import { injectable } from "inversify";
 import { AbstractInMemoryRepository } from "./abstract-in-memory-repository.js";
 import {ArticleRepository} from "../../repository/article-repository.js";
 import {FilterCondition} from "../../filter-condition.js";
+import {inMemoryWikiPageStore} from "./in-memory-wiki-page-store.js";
 
 @injectable()
 export class InMemoryArticleRepository
 	extends AbstractInMemoryRepository<Article>
 	implements ArticleRepository {
+
+	items = inMemoryWikiPageStore as Map<string, Article>;
 
 	findOneByNameAndWorld(name: string, worldId: string): Promise<Article> {
 		return this.findOne([