+++ packages/server/src/dal/in-memory/repositories/in-memory-game-repository.ts
@@ -2,16 +2,17 @@ import { Game } from "../../../domain-entities/game.js";
 import { injectable } from "inversify";
 import { AbstractInMemoryRepository } from "./abstract-in-memory-repository.js";
 import {GameRepository} from "../../repository/game-repository.js";
-import {FilterCondition} from "../../filter-condition.js";
 
 @injectable()
 export class InMemoryGameRepository extends AbstractInMemoryRepository<Game> implements GameRepository{
 
-    findWithModel(modelId: string): Promise<Game[]> {
-        return this.find([new FilterCondition("models", {_id: modelId})])
+    async findWithModel(modelId: string): Promise<Game[]> {
+        const games = await this.findAll();
+        return games.filter((game) => game.models.some((model) => model.model === modelId));
     }
 
-    findByPlayer(userId: string): Promise<Game[]> {
-        return this.find([new FilterCondition("characters.player", userId)]);
+    async findByPlayer(userId: string): Promise<Game[]> {
+        const games = await this.findAll();
+        return games.filter((game) => game.characters.some((character) => character.player === userId));
     }
 }