+++ packages/server/src/services/game-service.ts
@@ -12,7 +12,8 @@ import { SALT_ROUNDS } from "../resolvers/mutations/authentication-mutations.js"
 import { USER } from "@rpgtools/common/src/type-constants.js";
 import { SecurityContext } from "../security/security-context.js";
 import {
-	Character, CharacterAttribute,
+	Character,
+	CharacterAttribute,
 	Game,
 	InGameModel,
 	Message,